requirements.txt
requests==2.28.1
environs==9.5.0
python-telegram-bot==20.0a2
asyncio==3.4.3